chore: introduce a not found error
introduces a binding not found error that can be coded against when using GetComponentVersion.

This allows to check for existence of a component version
